erc1155 and erc20 decomposition
